My sales forecast for the Rossmann challenge
Link to Kaggle: https://www.kaggle.com/c/rossmann-store-sales/

In this notebook I start by exploring trends and features using pandas, matplotlib, and seaborn. After identifying the features and trends on which I want to focus, I filtered and prepared the data for linear regression.

I found that a 3rd-degree polynomial fitted the data quite well and thus fitted this function before running a linear regression. The final mean square error was 0.14936 when compared to the Kaggle submission dataset.